fix: type safety improvements across utilities and managers
- asset_manager: add config import for Settings, safe isoformat()
  on Optional datetime fields, fix undefined 'settings' reference
- character_manager: safe isoformat() on Optional created_at/updated_at
- game_ws: token param as Optional[str], player_role fallback
- r2_manager: type extra_args as dict[str, Any]
- roles: all functions accept Optional[str] for role parameters
- rate_limiter: window_minutes accepts float for sub-minute windows

+++ apps/server/utils/rate_limiter.py
@@ -11,7 +11,7 @@ class RateLimiter:
         # Store requests by IP address
         self.requests: Dict[str, Deque[float]] = defaultdict(deque)
         
-    def is_allowed(self, identifier: str, max_requests: int = 5, window_minutes: int = 5) -> bool:
+    def is_allowed(self, identifier: str, max_requests: int = 5, window_minutes: float = 5) -> bool:
         """
         Check if a request is allowed based on rate limiting.
         
@@ -41,7 +41,7 @@ class RateLimiter:
         
         return False
     
-    def get_time_until_reset(self, identifier: str, window_minutes: int = 5) -> int:
+    def get_time_until_reset(self, identifier: str, window_minutes: float = 5) -> int:
         """
         Get seconds until rate limit resets for this identifier.
